Workaround for missing PyGObject support for GStrv
Currently PyGObject doesn't support marshalling of string lists (see
https://bugzilla.gnome.org/show_bug.cgi?id=666636 ), so we need to
workaround this limitation somehow.
This is a quick hack: we add a GI overrides file which hooks into the
Signon.AuthSession.process() methods and converts string arrays into a
single concatenated string; this string gets recognized in
libsignon-glib and the argument is converted to a GStrv just before the
arguments are sent to D-Bus.